+++ src/app/classes/Express/Routes/Json.ts
@@ -1,8 +1,8 @@
-import log from '../lib/logger';
+import log from '../../../lib/logger';
 import express, { Router } from 'express';
 import SKU from '@tf2autobot/tf2-sku';
-import testSKU from '../lib/utils/testSKU';
-import Server from '../classes/Server';
+import testSKU from '../../../lib/utils/testSKU';
+import Server from '../../Server';
 import { Schema } from '@tf2autobot/tf2-schema';
 
 export default class Json {
